pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-image-gen"
version = "1.0.0"
description = "Interactive image generation using Google's Gemini API"
authors = [
    {name = "Gemini Image Gen Team"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "google",
    "google.genai",
    "gradio",
    "pillow",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
]

[project.scripts]
gemini-image-gen = "gemini_image_gen.ui.app:launch_app"

[project.urls]
Homepage = "https://github.com/yourusername/gemini-image-gen"
Documentation = "https://github.com/yourusername/gemini-image-gen/wiki"
Repository = "https://github.com/yourusername/gemini-image-gen.git"
Issues = "https://github.com/yourusername/gemini-image-gen/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gemini_image_gen*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=gemini_image_gen --cov-report=html --cov-report=term"